0.7.3
Added a .gitignore and worked on the shell patch(testing in miscfixes)

+++ DOW2DLLInjector/PlayerCFG.cpp
@@ -5,14 +5,14 @@ std::string cullexe(std::string path) {
 	int i = path.size() - 1;
 	for (i; i >= 0 && path[i] != '/'; i--);
 	std::string ret = "";
-	for (size_t j = 0; j <= i; j++) {
+	for (size_t j = 0; j <= (size_t)i; j++) {
 		ret.push_back(path[j]);
 	}
 	return ret;
 }
 //generates the path to cfg
 void Injector::createcfgpath(std::string path) {
-	for (int i = 0; i < path.size(); i++) {
+	for (int i = 0; (size_t)i < path.size(); i++) {
 		if (path[i] == '\\') {
 			path[i] = '/';
 		}
